python/requirements.txt
cocotb
pycryptodome
google-cloud-bigquery
dbt-bigquery
